requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
GPIOSimulator==0.1
pytz==2018.7
